requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2
